Dashboard UI - filter function removed temporarily
